requirements.txt
mplfinance>=0.12.10b0
pandas>=2.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
numpy>=1.24
pytest>=7.4
